update recycler in main activity

+++ app/src/main/java/com/example/weatherapplication/entities/WeatherTo5Days.kt
@@ -4,7 +4,7 @@ data class WeatherTo5Days(
     var newDay:String?=null,
     val time:Long,
     val text:String,
-    val temp:Int?,
+    val temp:Double?,
     val pressure: Int?,
     val wind: Double?
 )
